hotel_weather/requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
docutils==0.21.2
Geohash==1.0
idna==3.10
py4j==0.10.9.7
pyspark==3.5.4
python-dotenv==1.0.1
python-geohash==0.8.5
requests==2.32.3
urllib3==2.3.0
